ai-agent-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.5.0
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
textblob==0.17.1
nltk==3.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
asyncio==3.4.3
schedule==1.2.0
python-multipart==0.0.6
